+++ Client_Assembly/src/main/java/structs/User_Input.java
@@ -47,7 +47,7 @@ public class User_Input
     }
     public Object dyn_REG_get_Item_On_List_Of_InputSubsets(long praiseId)
     {
-        return stat_REG_get_List_Of_InputSubsets()[praiseId];
+        return stat_REG_get_List_Of_InputSubsets()[(int)praiseId];
     }
     // private.
     private static void stat_CLASS_boot0_DECLAIRE_User_Input()
